# docs/swagger.yaml
basePath: /api
definitions:
  models.Attendance:
    properties:
      checked_in_at:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      phone:
        type: string
      session:
        allOf:
        - $ref: '#/definitions/models.AttendanceSession'
        description: Relationships
      session_id:
        type: integer
      student_name:
        type: string
      updated_at:
        type: string
      work_unit:
        type: string
      work_unit_address:
        type: string
    type: object
  models.AttendanceSession:
    properties:
      attendances:
        items:
          $ref: '#/definitions/models.Attendance'
        type: array
      class:
        $ref: '#/definitions/models.Class'
      class_id:
        type: integer
      created_at:
        type: string
      event:
        allOf:
        - $ref: '#/definitions/models.Event'
        description: Relationships
      event_id:
        type: integer
      id:
        type: integer
      session_date:
        type: string
      teacher:
        $ref: '#/definitions/models.Teacher'
      teacher_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Class:
    properties:
      class_code:
        type: string
      class_name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      sessions:
        items:
          $ref: '#/definitions/models.AttendanceSession'
        type: array
      students:
        description: Relationships
        items:
          $ref: '#/definitions/models.Student'
        type: array
      updated_at:
        type: string
    type: object
...
  models.CreateEventRequest:
    properties:
      description:
        example: Workshop về trí tuệ nhân tạo
        type: string
      event_name:
        example: Workshop AI
        type: string
      start_date:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  models.CreateStudentRequest:
    properties:
      class_id:
        example: 1
        type: integer
      date_of_birth:
        example: "2000-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      phone:
        example: "0123456789"
        type: string
      student_code:
        example: SV001
        type: string
      student_name:
        example: Nguyen Van A
        type: string
      work_unit:
        example: Công ty ABC
        type: string
    type: object
  models.CreateTeacherRequest:
    properties:
      date_of_birth:
        example: "1980-01-01T00:00:00Z"
        type: string
      email:
        example: teacher@example.com
        type: string
      phone:
        example: "0123456789"
        type: string
      teacher_code:
        example: GV001
        type: string
      teacher_name:
        example: Nguyen Thi B
        type: string
      work_unit:
        example: Trường Đại học ABC
        type: string
    type: object
  models.Event:
    properties:
      created_at:
        type: string
      description:
        type: string
      event_name:
        type: string
      id:
        type: integer
      sessions:
        description: Relationships
        items:
          $ref: '#/definitions/models.AttendanceSession'
        type: array
      start_date:
        type: string
      updated_at:
        type: string
    type: object
  models.Student:
    properties:
      class:
        allOf:
        - $ref: '#/definitions/models.Class'
        description: Relationships
      class_id:
        type: integer
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      id:
        type: integer
      phone:
        type: string
      student_code:
        type: string
      student_name:
        type: string
      updated_at:
        type: string
      work_unit:
        type: string
    type: object
  models.Teacher:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      id:
        type: integer
      phone:
        type: string
      sessions:
        description: Relationships
        items:
          $ref: '#/definitions/models.AttendanceSession'
        type: array
      teacher_code:
        type: string
      teacher_name:
        type: string
      updated_at:
        type: string
      work_unit:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a simple student management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Student Management API
  version: "1.0"
paths:
...
  /events:
    get:
      consumes:
      - application/json
      description: Get a list of all events
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get all events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Create a new event with the provided information
      parameters:
      - description: Event data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new event
      tags:
      - events
  /events/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an event by its ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Delete an event
      tags:
      - events
    get:
      consumes:
      - application/json
      description: Get a single event by its ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get event by ID
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Update an existing event with the provided information
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Event data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.CreateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Update an event
      tags:
      - events
...
  /events/{id}/sessions:
    get:
      consumes:
      - application/json
      description: Get a single event by its ID including all attendance sessions
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "404":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get event with sessions
      tags:
      - events
...
  /students:
    get:
      consumes:
      - application/json
      description: Get a list of all students
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Student'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all students
      tags:
      - students
    post:
      consumes:
      - application/json
      description: Create a new student in the database
      parameters:
      - description: Student data
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.CreateStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new student
      tags:
      - students
  /teachers:
    get:
      description: Get all teachers from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Teacher'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all teachers
      tags:
      - teachers
    post:
      consumes:
      - application/json
      description: Create a new teacher in the database
      parameters:
      - description: Teacher data
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/models.CreateTeacherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Teacher'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new teacher
      tags:
      - teachers
  /teachers/{id}:
    get:
      description: Get a specific teacher by ID
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Teacher'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get teacher by ID
      tags:
      - teachers
swagger: "2.0"